requirement.txt
python-telegram-bot>=20.4
pandas>=2.0.3
matplotlib>=3.7.1
jieba>=0.42.1
wordcloud>=1.8.0.0